requirements.txt
apt-xapian-index==0.49
asgiref==3.4.1
attrs==21.2.0
backcall==0.2.0
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
click==8.0.1
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
decorator==5.0.9
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.2.5
djangorestframework==3.12.4
entrypoints==0.3
environs==9.3.2
Flask==2.0.0
gpg===1.13.1-unknown
httplib2==0.14.0
idna==2.8
iniconfig==1.1.1
ipdb==0.13.9
ipython==7.25.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==2.0.1
marshmallow==3.12.1
matplotlib-inline==0.1.2
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
packaging==20.9
parso==0.8.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
pluggy==0.13.1
prompt-toolkit==3.0.19
py==1.10.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.9.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.7
PyQt5==5.14.1
pytest==6.2.4
python-apt==2.0.0+ubuntu0.20.4.5
python-debian===0.1.36ubuntu1
python-dotenv==0.17.1
pytz==2021.1
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
sip==4.19.21
six==1.14.0
sqlparse==0.4.1
systemd-python==234
toml==0.10.2
traitlets==5.0.5
ubuntu-advantage-tools==27.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
wadllib==1.3.3
wcwidth==0.2.5
Werkzeug==2.0.1
xkit==0.0.0
